ISSUE #1219 move getSubModelRef to Ref
